% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxhull.R
\name{cxhullEdges}
\alias{cxhullEdges}
\title{Vertices and edges of convex hull}
\usage{
cxhullEdges(points, orderEdges = TRUE)
}
\arguments{
\item{points}{numeric matrix, one point per row}

\item{orderEdges}{Boolean, whether to order the edges in the output}
}
\value{
A list with two fields: \code{vertices} and \code{edges}.
}
\description{
Computes the vertices and the edges of the convex hull of a set 
  of points.
}
\examples{
library(cxhull)
}
