% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daVinciSphere}
\alias{daVinciSphere}
\title{Leonardo da Vinci's 72-sided sphere}
\format{
A matrix with 62 rows and 3 columns.
}
\source{
\url{http://www.matematicasvisuales.com/english/html/geometry/space/sphereCampanus.html}
}
\usage{
daVinciSphere
}
\description{
A matrix giving the 62 vertices of da Vinci's 72-sided sphere, 
  a convex polyhedra with 72 faces.
}
\keyword{datasets}
