% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxhull.R
\name{EdgesXYZ}
\alias{EdgesXYZ}
\title{Edges coordinates}
\usage{
EdgesXYZ(hull)
}
\arguments{
\item{hull}{an output of \code{\link{cxhull}} applied to 3D points and 
with the option \code{triangulate=TRUE}}
}
\value{
A numeric matrix with four columns. The first three values of a row 
  are the coordinates of a vertex at the extremity of an edge, and the 
  fourth column indicates whether the edge is a border edge.
}
\description{
The coordinates of the extremities of the edges in a matrix, 
  plus a column indicating which edges are border edges.
}
