% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hexacosichoron}
\alias{hexacosichoron}
\title{Vertices of the 600-cell}
\format{
A matrix with 120 rows and 4 columns.
}
\source{
\url{http://eusebeia.dyndns.org/4d/600-cell/}
}
\usage{
hexacosichoron
}
\description{
A matrix giving the 120 vertices of the hexacosichoron, a 
  regular 4D polytope also known as the "600-cell", with edge length 
  \code{2/phi}, where \code{phi} is the golden number. It has 720 edges.
}
\keyword{datasets}
