% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxhull.R
\name{VerticesXYZ}
\alias{VerticesXYZ}
\title{Convex hull vertices}
\usage{
VerticesXYZ(hull)
}
\arguments{
\item{hull}{an output of \code{\link{cxhull}} applied to 3D points}
}
\value{
A matrix with three columns. Each row represents the coordinates of 
  a vertex and the row names are the ids of the vertices.
}
\description{
The coordinates of the vertices of a 3D convex hull.
}
